将mq替换为http方式主动请求

+++ rmq-service/src/main/java/com/cn/rmq/service/impl/RmqServiceImpl.java
@@ -1,19 +1,23 @@
 package com.cn.rmq.service.impl;
 
 import cn.hutool.core.util.IdUtil;
+import cn.hutool.http.HttpException;
+import cn.hutool.http.HttpUtil;
+import cn.hutool.json.JSONObject;
+import cn.hutool.json.JSONUtil;
+import com.alibaba.dubbo.config.annotation.Reference;
 import com.alibaba.dubbo.config.annotation.Service;
 import com.cn.rmq.api.enums.MessageStatusEnum;
 import com.cn.rmq.api.exceptions.CheckException;
 import com.cn.rmq.api.model.Constants;
 import com.cn.rmq.api.model.RmqMessage;
 import com.cn.rmq.api.model.po.Message;
+import com.cn.rmq.api.model.po.Queue;
+import com.cn.rmq.api.service.IMessageService;
 import com.cn.rmq.api.service.IRmqService;
-import com.cn.rmq.api.service.User;
 import com.cn.rmq.dal.mapper.MessageMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jms.core.JmsMessagingTemplate;
 
 import java.time.LocalDateTime;
 import java.util.Arrays;
@@ -30,21 +34,15 @@ import java.util.List;
 public class RmqServiceImpl extends BaseServiceImpl<MessageMapper, Message, String>
         implements IRmqService {
 
-    @Autowired
-    private JmsMessagingTemplate jmsMessagingTemplate;
-
+    @Reference
+    private IMessageService messageService;
 
     @Override
-    public List<User> getUsers() {
-        System.out.println(">>> Client address from http servlet request: " );
-        return Arrays.asList(new User(1L, "User1"), new User(2L, "User2"));
-    }
-    @Override
-    public String createPreMessage(String consumerQueue,String messageBody) {
-        if (StringUtils.isBlank(consumerQueue)) {
+    public RmqMessage createPreMessage(RmqMessage rmqMessage) {
+        if (StringUtils.isBlank(rmqMessage.getConsumerQueue())) {
             throw new CheckException("consumerQueue is empty");
         }
-        if (StringUtils.isBlank(messageBody)) {
+        if (StringUtils.isBlank(rmqMessage.getMessageBody())) {
             throw new CheckException("messageBody is empty");
         }
 
@@ -53,17 +51,21 @@ public class RmqServiceImpl extends BaseServiceImpl<MessageMapper, Message, Stri
         // 插入预发送消息记录
         Message message = new Message();
         message.setId(id);
-        message.setConsumerQueue(consumerQueue);
-        message.setMessageBody(messageBody);
+        message.setConsumerQueue(rmqMessage.getConsumerQueue());
+        message.setMessageBody(rmqMessage.getMessageBody());
         message.setCreateTime(LocalDateTime.now());
         message.setUpdateTime(LocalDateTime.now());
         mapper.insertSelective(message);
 
-        return id;
+        rmqMessage.setConsumerQueue(null);
+        rmqMessage.setMessageBody(null);
+        rmqMessage.setMessageId(id);
+        return rmqMessage;
     }
 
     @Override
-    public void confirmAndSendMessage(String messageId) {
+    public void confirmAndSendMessage(RmqMessage rmqMessage) {
+        String messageId = rmqMessage.getMessageId();
         if (StringUtils.isBlank(messageId)) {
             throw new CheckException("messageId is empty");
         }
@@ -82,19 +84,18 @@ public class RmqServiceImpl extends BaseServiceImpl<MessageMapper, Message, Stri
         updateBean.setUpdateTime(LocalDateTime.now());
         mapper.updateByPrimaryKeySelective(updateBean);
 
-        // 发送MQ消息
-        RmqMessage rmqMessage = new RmqMessage();
-        rmqMessage.setMessageId(messageId);
+
         rmqMessage.setMessageBody(message.getMessageBody());
-        jmsMessagingTemplate.convertAndSend(message.getConsumerQueue(), rmqMessage);
+        // 发送http消息
+        postMessage(rmqMessage);
     }
 
     @Override
-    public void saveAndSendMessage(String consumerQueue, String messageBody) {
-        if (StringUtils.isBlank(consumerQueue)) {
+    public void saveAndSendMessage(RmqMessage rmqMessage) {
+        if (StringUtils.isBlank(rmqMessage.getConsumerQueue())) {
             throw new CheckException("consumerQueue is empty");
         }
-        if (StringUtils.isBlank(messageBody)) {
+        if (StringUtils.isBlank(rmqMessage.getMessageBody())) {
             throw new CheckException("messageBody is empty");
         }
 
@@ -103,23 +104,23 @@ public class RmqServiceImpl extends BaseServiceImpl<MessageMapper, Message, Stri
         // 插入发送消息记录
         Message message = new Message();
         message.setId(id);
-        message.setConsumerQueue(consumerQueue);
-        message.setMessageBody(messageBody);
+        message.setConsumerQueue(rmqMessage.getConsumerQueue());
+        message.setMessageBody(rmqMessage.getMessageBody());
         message.setStatus(MessageStatusEnum.SENDING.getValue());
         message.setCreateTime(LocalDateTime.now());
         message.setConfirmTime(LocalDateTime.now());
         message.setUpdateTime(LocalDateTime.now());
         mapper.insertSelective(message);
 
-        // 发送MQ消息
-        RmqMessage rmqMessage = new RmqMessage();
-        rmqMessage.setMessageId(id);
-        rmqMessage.setMessageBody(messageBody);
-        jmsMessagingTemplate.convertAndSend(consumerQueue, rmqMessage);
+        // 发送http消息
+        postMessage(rmqMessage);
     }
 
     @Override
-    public void directSendMessage(String consumerQueue, String messageBody) {
+    public void directSendMessage(RmqMessage rmqMessage) {
+        String messageBody = rmqMessage.getMessageBody();
+        String consumerQueue =rmqMessage.getConsumerQueue();
+
         if (StringUtils.isBlank(consumerQueue)) {
             throw new CheckException("consumerQueue is empty");
         }
@@ -127,14 +128,60 @@ public class RmqServiceImpl extends BaseServiceImpl<MessageMapper, Message, Stri
             throw new CheckException("messageBody is empty");
         }
 
-        // 发送MQ消息
-        RmqMessage rmqMessage = new RmqMessage();
-        rmqMessage.setMessageBody(messageBody);
-        jmsMessagingTemplate.convertAndSend(consumerQueue, rmqMessage);
+        // 发送http消息
+        postMessage(rmqMessage);
     }
 
     @Override
-    public void deleteMessageById(String messageId) {
+    public void deleteMessageById(RmqMessage rmqMessage) {
+        String messageId = rmqMessage.getMessageId();
+        if (StringUtils.isBlank(messageId)) {
+            throw new CheckException("messageId is empty");
+        }
         mapper.deleteByPrimaryKey(messageId);
+    }
+
+
+
+    @Override
+    public void postMessage(RmqMessage rmqMessage) {
+        try {
+            String messageBody = rmqMessage.getMessageBody();
+            Short postTimeout = rmqMessage.getPostTimeout();
+            log.info("投递消息=[{}]", JSONUtil.toJsonStr(messageBody));
+            // 调用下层业务方http接口投递消息
+            String postRsp = HttpUtil.post(rmqMessage.getPostUrl(), messageBody, postTimeout);
+            log.info("投递消息id=[{}], 返回应答=[{}]", rmqMessage.getMessageId(), postRsp);
+
+            // 解析业务方返回结果
+            JSONObject jsonObject = JSONUtil.parseObj(postRsp);
+            Integer code = jsonObject.getInt(Constants.KEY_CODE);
+            if (code.equals(Constants.CODE_SUCCESS)) {
+                // code=CODE_SUCCESS，业务方处理正常
+                Integer data = jsonObject.getInt(Constants.KEY_DATA);
+                if (data == 1) {
+                    // data=1，该消息业务处理成功
+                    log.info("下层处理消息[{}]成功",rmqMessage.getMessageId());
+
+                    deleteMessageById(rmqMessage);
+                } else {
+                    // data!=1，该消息业务处理失败，标记为死亡，不继续发送
+                    log.info("下层处理消息[{}]失败，标记为死亡",rmqMessage.getMessageId());
+
+                    //不考虑发送次数，直接标记为已死亡
+                    int updateCount = messageService.updateMessageDeadById(rmqMessage.getMessageId());
+                    log.info("更新[{}]条记录", updateCount);
+
+                }
+            } else {
+                // 业务方处理异常，记录日志---->系统异常
+                String msg = jsonObject.getStr(Constants.KEY_MSG);
+                log.error("消息投递失败, messageId=[{}], code=[{}], msg=[{}]", rmqMessage.getMessageId(), code, msg);
+            }
+        } catch (HttpException e) {
+            log.error("消息投递失败, messageId=[" + rmqMessage.getMessageId() + "], error:", e);
+        } catch (Exception e) {
+            log.error("消息投递失败, messageId=[" + rmqMessage.getMessageId() + "], error:", e);
+        }
     }
 }
